Split QgsBrowserDockWidget to new QgsBrowserWidget class and QgsBrowserDockWidget
Moves the guts of QgsBrowserDockWidget to a reusable QWidget subclass
"QgsBrowserWidget", so that browser widgets can easily be created
which aren't a dock widget.
