+++ src/gui/qgsbrowserdockwidget.cpp
@@ -14,272 +14,52 @@
  ***************************************************************************/
 #include "qgsbrowserdockwidget.h"
 #include "qgsbrowserdockwidget_p.h"
-
-#include <QAbstractTextDocumentLayout>
-#include <QHeaderView>
-#include <QTreeView>
-#include <QMenu>
-#include <QToolButton>
-#include <QFileDialog>
-#include <QPlainTextDocumentLayout>
-#include <QSortFilterProxyModel>
-
-#include "qgsbrowserguimodel.h"
-#include "qgsbrowsertreeview.h"
-#include "qgslogger.h"
-#include "qgsrasterlayer.h"
-#include "qgsvectorlayer.h"
-#include "qgsproject.h"
-#include "qgssettings.h"
-#include "qgsnewnamedialog.h"
+#include "qgsbrowserwidget.h"
 #include "qgsbrowserproxymodel.h"
-#include "qgsgui.h"
-#include "qgswindowmanagerinterface.h"
-#include "qgsnative.h"
-#include "qgsdataitemguiproviderregistry.h"
-#include "qgsdataitemguiprovider.h"
+#include "qgsbrowserguimodel.h"
 #include "qgsdirectoryitem.h"
-#include "qgslayeritem.h"
 #include "qgsprojectitem.h"
+#include "qgslayeritem.h"
 
-// browser layer properties dialog
-#include "qgsapplication.h"
-#include "qgsmapcanvas.h"
-
-#include <QDragEnterEvent>
-#include <functional>
+#include <QVBoxLayout>
+#include <QFileDialog>
 
 QgsBrowserDockWidget::QgsBrowserDockWidget( const QString &name, QgsBrowserGuiModel *browserModel, QWidget *parent )
   : QgsDockWidget( parent )
-  , mModel( browserModel )
-  , mPropertiesWidgetEnabled( false )
-  , mPropertiesWidgetHeight( 0 )
 {
-  setupUi( this );
-
-  mContents->layout()->setContentsMargins( 0, 0, 0, 0 );
-  static_cast< QVBoxLayout * >( mContents->layout() )->setSpacing( 0 );
+  QVBoxLayout *layout = new QVBoxLayout();
+  layout->setContentsMargins( 0, 0, 0, 0 );
+  layout->setSpacing( 0 );
+  QWidget *container = new QWidget();
+  container->setLayout( layout );
+  setWidget( container );
 
   setWindowTitle( name );
 
-  mBrowserView = new QgsDockBrowserTreeView( this );
-  mLayoutBrowser->addWidget( mBrowserView );
+  mWidget = new QgsBrowserWidget( browserModel );
+  layout->addWidget( mWidget );
 
-  mWidgetFilter->hide();
-  mLeFilter->setPlaceholderText( tr( "Type here to filter visible items…" ) );
-  // icons from http://www.fatcow.com/free-icons License: CC Attribution 3.0
-
-  QMenu *menu = new QMenu( this );
-  menu->setSeparatorsCollapsible( false );
-  mBtnFilterOptions->setMenu( menu );
-  QAction *action = new QAction( tr( "Case Sensitive" ), menu );
-  action->setData( "case" );
-  action->setCheckable( true );
-  action->setChecked( false );
-  connect( action, &QAction::toggled, this, &QgsBrowserDockWidget::setCaseSensitive );
-  menu->addAction( action );
-  QActionGroup *group = new QActionGroup( menu );
-  action = new QAction( tr( "Filter Pattern Syntax" ), group );
-  action->setSeparator( true );
-  menu->addAction( action );
-  action = new QAction( tr( "Normal" ), group );
-  action->setData( QgsBrowserProxyModel::Normal );
-  action->setCheckable( true );
-  action->setChecked( true );
-  menu->addAction( action );
-  action = new QAction( tr( "Wildcard(s)" ), group );
-  action->setData( QgsBrowserProxyModel::Wildcards );
-  action->setCheckable( true );
-  menu->addAction( action );
-  action = new QAction( tr( "Regular Expression" ), group );
-  action->setData( QgsBrowserProxyModel::RegularExpression );
-  action->setCheckable( true );
-  menu->addAction( action );
-
-  mBrowserView->setExpandsOnDoubleClick( false );
-
-  connect( mActionRefresh, &QAction::triggered, this, &QgsBrowserDockWidget::refresh );
-  connect( mActionAddLayers, &QAction::triggered, this, &QgsBrowserDockWidget::addSelectedLayers );
-  connect( mActionCollapse, &QAction::triggered, mBrowserView, &QgsDockBrowserTreeView::collapseAll );
-  connect( mActionShowFilter, &QAction::triggered, this, &QgsBrowserDockWidget::showFilterWidget );
-  connect( mActionPropertiesWidget, &QAction::triggered, this, &QgsBrowserDockWidget::enablePropertiesWidget );
-  connect( mLeFilter, &QgsFilterLineEdit::returnPressed, this, &QgsBrowserDockWidget::setFilter );
-  connect( mLeFilter, &QgsFilterLineEdit::cleared, this, &QgsBrowserDockWidget::setFilter );
-  connect( mLeFilter, &QgsFilterLineEdit::textChanged, this, &QgsBrowserDockWidget::setFilter );
-  connect( group, &QActionGroup::triggered, this, &QgsBrowserDockWidget::setFilterSyntax );
-  connect( mBrowserView, &QgsDockBrowserTreeView::customContextMenuRequested, this, &QgsBrowserDockWidget::showContextMenu );
-  connect( mBrowserView, &QgsDockBrowserTreeView::doubleClicked, this, &QgsBrowserDockWidget::itemDoubleClicked );
-  connect( mSplitter, &QSplitter::splitterMoved, this, &QgsBrowserDockWidget::splitterMoved );
-
-  connect( QgsGui::instance(), &QgsGui::optionsChanged, this, &QgsBrowserDockWidget::onOptionsChanged );
+  connect( mWidget, &QgsBrowserWidget::openFile, this, &QgsBrowserDockWidget::openFile );
+  connect( mWidget, &QgsBrowserWidget::handleDropUriList, this, &QgsBrowserDockWidget::handleDropUriList );
+  connect( mWidget, &QgsBrowserWidget::connectionsChanged, this, &QgsBrowserDockWidget::connectionsChanged );
 }
 
-QgsBrowserDockWidget::~QgsBrowserDockWidget()
+QgsBrowserDockWidget::~QgsBrowserDockWidget() = default;
+
+QgsBrowserWidget *QgsBrowserDockWidget::browserWidget()
 {
-  QgsSettings settings;
-  settings.setValue( settingsSection() + "/propertiesWidgetEnabled", mPropertiesWidgetEnabled );
-  //settings.setValue(settingsSection() + "/propertiesWidgetHeight", mPropertiesWidget->size().height() );
-  settings.setValue( settingsSection() + "/propertiesWidgetHeight", mPropertiesWidgetHeight );
-}
-
-void QgsBrowserDockWidget::showEvent( QShowEvent *e )
-{
-  // delayed initialization of the model
-  if ( !mModel->initialized( ) )
-  {
-    mModel->initialize();
-  }
-  if ( ! mProxyModel )
-  {
-    mProxyModel = new QgsBrowserProxyModel( this );
-    mProxyModel->setBrowserModel( mModel );
-    mProxyModel->setHiddenDataItemProviderKeyFilter( mDisabledDataItemsKeys );
-    mBrowserView->setSettingsSection( objectName().toLower() ); // to distinguish 2 or more instances of the browser
-    mBrowserView->setBrowserModel( mModel );
-    mBrowserView->setModel( mProxyModel );
-    mBrowserView->setSortingEnabled( true );
-    mBrowserView->sortByColumn( 0, Qt::AscendingOrder );
-    // provide a horizontal scroll bar instead of using ellipse (...) for longer items
-    mBrowserView->setTextElideMode( Qt::ElideNone );
-    mBrowserView->header()->setSectionResizeMode( 0, QHeaderView::ResizeToContents );
-    mBrowserView->header()->setStretchLastSection( false );
-
-    // selectionModel is created when model is set on tree
-    connect( mBrowserView->selectionModel(), &QItemSelectionModel::selectionChanged,
-             this, &QgsBrowserDockWidget::selectionChanged );
-
-    // Forward the model changed signals to the widget
-    connect( mModel, &QgsBrowserModel::connectionsChanged,
-             this, &QgsBrowserDockWidget::connectionsChanged );
-
-
-    // objectName used by settingsSection() is not yet set in constructor
-    QgsSettings settings;
-    mPropertiesWidgetEnabled = settings.value( settingsSection() + "/propertiesWidgetEnabled", false ).toBool();
-    mActionPropertiesWidget->setChecked( mPropertiesWidgetEnabled );
-    mPropertiesWidget->setVisible( false ); // false until item is selected
-
-    mPropertiesWidgetHeight = settings.value( settingsSection() + "/propertiesWidgetHeight" ).toFloat();
-    QList<int> sizes = mSplitter->sizes();
-    int total = sizes.value( 0 ) + sizes.value( 1 );
-    int height = static_cast<int>( total * mPropertiesWidgetHeight );
-    sizes.clear();
-    sizes << total - height << height;
-    mSplitter->setSizes( sizes );
-  }
-
-  QgsDockWidget::showEvent( e );
-}
-
-void QgsBrowserDockWidget::itemDoubleClicked( const QModelIndex &index )
-{
-  QgsDataItem *item = mModel->dataItem( mProxyModel->mapToSource( index ) );
-  if ( !item )
-    return;
-
-  QgsDataItemGuiContext context = createContext();
-
-  const QList< QgsDataItemGuiProvider * > providers = QgsGui::instance()->dataItemGuiProviderRegistry()->providers();
-  for ( QgsDataItemGuiProvider *provider : providers )
-  {
-    if ( provider->handleDoubleClick( item, context ) )
-      return;
-  }
-
-  // if no providers overrode the double-click handling for this item, we give the item itself a chance
-  if ( !item->handleDoubleClick() )
-  {
-    // double-click not handled by browser model, so use as default view expand behavior
-    if ( mBrowserView->isExpanded( index ) )
-      mBrowserView->collapse( index );
-    else
-      mBrowserView->expand( index );
-  }
-}
-
-void QgsBrowserDockWidget::onOptionsChanged()
-{
-  std::function< void( const QModelIndex &index ) > updateItem;
-  updateItem = [this, &updateItem]( const QModelIndex & index )
-  {
-    if ( QgsDirectoryItem *dirItem = qobject_cast< QgsDirectoryItem * >( mModel->dataItem( index ) ) )
-    {
-      dirItem->reevaluateMonitoring();
-    }
-
-    const int rowCount = mModel->rowCount( index );
-    for ( int i = 0; i < rowCount; ++i )
-    {
-      const QModelIndex child = mModel->index( i, 0, index );
-      updateItem( child );
-    }
-  };
-
-  for ( int i = 0; i < mModel->rowCount(); ++i )
-  {
-    updateItem( mModel->index( i, 0 ) );
-  }
+  return mWidget;
 }
 
 void QgsBrowserDockWidget::showContextMenu( QPoint pt )
 {
-  QModelIndex index = mProxyModel->mapToSource( mBrowserView->indexAt( pt ) );
-  QgsDataItem *item = mModel->dataItem( index );
-  if ( !item )
-    return;
-
-  const QModelIndexList selection = mBrowserView->selectionModel()->selectedIndexes();
-  QList< QgsDataItem * > selectedItems;
-  selectedItems.reserve( selection.size() );
-  for ( const QModelIndex &selectedIndex : selection )
-  {
-    QgsDataItem *selectedItem = mProxyModel->dataItem( selectedIndex );
-    if ( selectedItem )
-      selectedItems << selectedItem;
-  }
-
-  QMenu *menu = new QMenu( this );
-
-  const QList<QMenu *> menus = item->menus( menu );
-  QList<QAction *> actions = item->actions( menu );
-
-  if ( !menus.isEmpty() )
-  {
-    for ( QMenu *mn : menus )
-    {
-      menu->addMenu( mn );
-    }
-  }
-
-  if ( !actions.isEmpty() )
-  {
-    if ( !menu->actions().isEmpty() )
-      menu->addSeparator();
-    // add action to the menu
-    menu->addActions( actions );
-  }
-
-  QgsDataItemGuiContext context = createContext();
-
-  const QList< QgsDataItemGuiProvider * > providers = QgsGui::instance()->dataItemGuiProviderRegistry()->providers();
-  for ( QgsDataItemGuiProvider *provider : providers )
-  {
-    provider->populateContextMenu( item, menu, selectedItems, context );
-  }
-
-  if ( menu->actions().isEmpty() )
-  {
-    delete menu;
-    return;
-  }
-
-  menu->popup( mBrowserView->mapToGlobal( pt ) );
+  mWidget->showContextMenu( pt );
 }
 
 void QgsBrowserDockWidget::addFavorite()
 {
-  QModelIndex index = mProxyModel->mapToSource( mBrowserView->currentIndex() );
-  QgsDataItem *item = mModel->dataItem( index );
+  QModelIndex index = mWidget->mProxyModel->mapToSource( mWidget->mBrowserView->currentIndex() );
+  QgsDataItem *item = mWidget->mModel->dataItem( index );
   if ( !item )
     return;
 
@@ -305,94 +85,38 @@ void QgsBrowserDockWidget::addFavoriteDirectory()
 
 void QgsBrowserDockWidget::addFavoriteDirectory( const QString &favDir, const QString &name )
 {
-  mModel->addFavoriteDirectory( favDir, name );
+  mWidget->mModel->addFavoriteDirectory( favDir, name );
 }
 
 void QgsBrowserDockWidget::setMessageBar( QgsMessageBar *bar )
 {
-  mMessageBar = bar;
-  mModel->setMessageBar( bar );
+  mWidget->setMessageBar( bar );
 }
 
 QgsMessageBar *QgsBrowserDockWidget::messageBar()
 {
-  return mMessageBar;
+  return mWidget->messageBar();
 }
 
 void QgsBrowserDockWidget::setDisabledDataItemsKeys( const QStringList &filter )
 {
-  mDisabledDataItemsKeys = filter;
-
-  if ( !mProxyModel )
-    return;
-
-  mProxyModel->setHiddenDataItemProviderKeyFilter( mDisabledDataItemsKeys );
+  mWidget->setDisabledDataItemsKeys( filter );
 }
 
 void QgsBrowserDockWidget::removeFavorite()
 {
-  mModel->removeFavorite( mProxyModel->mapToSource( mBrowserView->currentIndex() ) );
+  mWidget->mModel->removeFavorite( mWidget->mProxyModel->mapToSource( mWidget->mBrowserView->currentIndex() ) );
 }
 
 void QgsBrowserDockWidget::refresh()
 {
-  refreshModel( QModelIndex() );
-}
-
-void QgsBrowserDockWidget::refreshModel( const QModelIndex &index )
-{
-  if ( mModel && mProxyModel )
-  {
-    QgsDataItem *item = mModel->dataItem( index );
-    if ( item )
-    {
-      QgsDebugMsgLevel( "path = " + item->path(), 4 );
-    }
-    else
-    {
-      QgsDebugMsgLevel( QStringLiteral( "invalid item" ), 4 );
-    }
-
-    if ( item && ( item->capabilities2() & Qgis::BrowserItemCapability::Fertile ) )
-    {
-      mModel->refresh( index );
-    }
-
-    for ( int i = 0; i < mModel->rowCount( index ); i++ )
-    {
-      QModelIndex idx = mModel->index( i, 0, index );
-      QModelIndex proxyIdx = mProxyModel->mapFromSource( idx );
-      QgsDataItem *child = mModel->dataItem( idx );
-
-      // Check also expanded descendants so that the whole expanded path does not get collapsed if one item is collapsed.
-      // Fast items (usually root items) are refreshed so that when collapsed, it is obvious they are if empty (no expand symbol).
-      if ( mBrowserView->isExpanded( proxyIdx ) || mBrowserView->hasExpandedDescendant( proxyIdx ) || ( child && child->capabilities2() & Qgis::BrowserItemCapability::Fast ) )
-      {
-        refreshModel( idx );
-      }
-      else
-      {
-        if ( child && ( child->capabilities2() & Qgis::BrowserItemCapability::Fertile ) )
-        {
-          child->depopulate();
-        }
-      }
-    }
-  }
-}
-
-void QgsBrowserDockWidget::addLayer( QgsLayerItem *layerItem )
-{
-  if ( !layerItem )
-    return;
-
-  emit handleDropUriList( layerItem->mimeUris() );
+  mWidget->refreshModel( QModelIndex() );
 }
 
 bool QgsBrowserDockWidget::addLayerAtIndex( const QModelIndex &index )
 {
-  QgsDebugMsg( QStringLiteral( "rowCount() = %1" ).arg( mModel->rowCount( mProxyModel->mapToSource( index ) ) ) );
-  QgsDataItem *item = mModel->dataItem( mProxyModel->mapToSource( index ) );
+  QgsDebugMsg( QStringLiteral( "rowCount() = %1" ).arg( mWidget->mModel->rowCount( mWidget->mProxyModel->mapToSource( index ) ) ) );
+  QgsDataItem *item = mWidget->mModel->dataItem( mWidget->mProxyModel->mapToSource( index ) );
 
   if ( item && item->type() == Qgis::BrowserItemType::Project )
   {
@@ -411,7 +135,7 @@ bool QgsBrowserDockWidget::addLayerAtIndex( const QModelIndex &index )
     if ( layerItem )
     {
       QApplication::setOverrideCursor( Qt::WaitCursor );
-      addLayer( layerItem );
+      mWidget->addLayer( layerItem );
       QApplication::restoreOverrideCursor();
     }
     return true;
@@ -421,76 +145,23 @@ bool QgsBrowserDockWidget::addLayerAtIndex( const QModelIndex &index )
 
 void QgsBrowserDockWidget::addSelectedLayers()
 {
-  QApplication::setOverrideCursor( Qt::WaitCursor );
-
-  // get a sorted list of selected indexes
-  QModelIndexList list = mBrowserView->selectionModel()->selectedIndexes();
-  std::sort( list.begin(), list.end() );
-
-  // If any of the layer items are QGIS we just open and exit the loop
-  const auto constList = list;
-  for ( const QModelIndex &index : constList )
-  {
-    QgsDataItem *item = mModel->dataItem( mProxyModel->mapToSource( index ) );
-    if ( item && item->type() == Qgis::BrowserItemType::Project )
-    {
-      QgsProjectItem *projectItem = qobject_cast<QgsProjectItem *>( item );
-      if ( projectItem )
-        emit openFile( projectItem->path(), QStringLiteral( "project" ) );
-
-      QApplication::restoreOverrideCursor();
-      return;
-    }
-  }
-
-  // add items in reverse order so they are in correct order in the layers dock
-  for ( int i = list.size() - 1; i >= 0; i-- )
-  {
-    QgsDataItem *item = mModel->dataItem( mProxyModel->mapToSource( list[i] ) );
-    if ( item && item->type() == Qgis::BrowserItemType::Layer )
-    {
-      QgsLayerItem *layerItem = qobject_cast<QgsLayerItem *>( item );
-      if ( layerItem )
-        addLayer( layerItem );
-    }
-  }
-
-  QApplication::restoreOverrideCursor();
+  mWidget->addSelectedLayers();
 }
 
 void QgsBrowserDockWidget::hideItem()
 {
-  QModelIndex index = mProxyModel->mapToSource( mBrowserView->currentIndex() );
-  QgsDataItem *item = mModel->dataItem( index );
-  if ( ! item )
-    return;
-
-  if ( item->type() == Qgis::BrowserItemType::Directory )
-  {
-    mModel->hidePath( item );
-  }
+  mWidget->hideItem();
 }
 
 void QgsBrowserDockWidget::showProperties()
 {
-  QModelIndex index = mProxyModel->mapToSource( mBrowserView->currentIndex() );
-  QgsDataItem *item = mModel->dataItem( index );
-  if ( ! item )
-    return;
-
-  if ( item->type() == Qgis::BrowserItemType::Layer || item->type() == Qgis::BrowserItemType::Directory )
-  {
-    QgsBrowserPropertiesDialog *dialog = new QgsBrowserPropertiesDialog( settingsSection(), this );
-    dialog->setAttribute( Qt::WA_DeleteOnClose );
-    dialog->setItem( item, createContext() );
-    dialog->show();
-  }
+  mWidget->showProperties();
 }
 
 void QgsBrowserDockWidget::toggleFastScan()
 {
-  QModelIndex index = mProxyModel->mapToSource( mBrowserView->currentIndex() );
-  QgsDataItem *item = mModel->dataItem( index );
+  QModelIndex index = mWidget->mProxyModel->mapToSource( mWidget->mBrowserView->currentIndex() );
+  QgsDataItem *item = mWidget->mModel->dataItem( index );
   if ( ! item )
     return;
 
@@ -514,131 +185,44 @@ void QgsBrowserDockWidget::toggleFastScan()
 
 void QgsBrowserDockWidget::showFilterWidget( bool visible )
 {
-  mWidgetFilter->setVisible( visible );
-  if ( ! visible )
-  {
-    mLeFilter->setText( QString() );
-    setFilter();
-  }
-  else
-  {
-    mLeFilter->setFocus();
-  }
+  mWidget->showFilterWidget( visible );
 }
 
 void QgsBrowserDockWidget::setFilter()
 {
-  QString filter = mLeFilter->text();
-  if ( mProxyModel )
-    mProxyModel->setFilterString( filter );
+  mWidget->setFilter();
 }
 
 void QgsBrowserDockWidget::updateProjectHome()
 {
-  if ( mModel )
-    mModel->updateProjectHome();
+  mWidget->updateProjectHome();
 }
 
 void QgsBrowserDockWidget::setFilterSyntax( QAction *action )
 {
-  if ( !action || ! mProxyModel )
-    return;
-
-  mProxyModel->setFilterSyntax( static_cast< QgsBrowserProxyModel::FilterSyntax >( action->data().toInt() ) );
+  mWidget->setFilterSyntax( action );
 }
 
 void QgsBrowserDockWidget::setCaseSensitive( bool caseSensitive )
 {
-  if ( ! mProxyModel )
-    return;
-  mProxyModel->setFilterCaseSensitivity( caseSensitive ? Qt::CaseSensitive : Qt::CaseInsensitive );
-}
-
-int QgsBrowserDockWidget::selectedItemsCount()
-{
-  QItemSelectionModel *selectionModel = mBrowserView->selectionModel();
-  if ( selectionModel )
-  {
-    return selectionModel->selectedIndexes().size();
-  }
-  return 0;
-}
-
-QgsDataItemGuiContext QgsBrowserDockWidget::createContext()
-{
-  QgsDataItemGuiContext context;
-  context.setMessageBar( mMessageBar );
-  return context;
+  mWidget->setCaseSensitive( caseSensitive );
 }
 
 void QgsBrowserDockWidget::selectionChanged( const QItemSelection &selected, const QItemSelection &deselected )
 {
-  Q_UNUSED( selected )
-  Q_UNUSED( deselected )
-  if ( mPropertiesWidgetEnabled )
-  {
-    setPropertiesWidget();
-  }
-}
-
-void QgsBrowserDockWidget::clearPropertiesWidget()
-{
-  while ( mPropertiesLayout->count() > 0 )
-  {
-    delete mPropertiesLayout->itemAt( 0 )->widget();
-  }
-  mPropertiesWidget->setVisible( false );
-}
-
-void QgsBrowserDockWidget::setPropertiesWidget()
-{
-  clearPropertiesWidget();
-  QItemSelectionModel *selectionModel = mBrowserView->selectionModel();
-  if ( selectionModel )
-  {
-    QModelIndexList indexes = selectionModel->selectedIndexes();
-    if ( indexes.size() == 1 )
-    {
-      QModelIndex index = mProxyModel->mapToSource( indexes.value( 0 ) );
-      QgsDataItem *item = mModel->dataItem( index );
-      QgsDataItemGuiContext context = createContext();
-      QgsBrowserPropertiesWidget *propertiesWidget = QgsBrowserPropertiesWidget::createWidget( item, context, mPropertiesWidget );
-      if ( propertiesWidget )
-      {
-        propertiesWidget->setCondensedMode( true );
-        mPropertiesLayout->addWidget( propertiesWidget );
-      }
-    }
-  }
-  mPropertiesWidget->setVisible( mPropertiesLayout->count() > 0 );
+  mWidget->selectionChanged( selected, deselected );
 }
 
 void QgsBrowserDockWidget::enablePropertiesWidget( bool enable )
 {
-  mPropertiesWidgetEnabled = enable;
-  if ( enable && selectedItemsCount() == 1 )
-  {
-    setPropertiesWidget();
-  }
-  else
-  {
-    clearPropertiesWidget();
-  }
+  mWidget->enablePropertiesWidget( enable );
 }
 
 void QgsBrowserDockWidget::setActiveIndex( const QModelIndex &index )
 {
-  if ( index.isValid() )
-  {
-    QModelIndex proxyIndex = mProxyModel->mapFromSource( index );
-    mBrowserView->expand( proxyIndex );
-    mBrowserView->setCurrentIndex( proxyIndex );
-  }
+  mWidget->setActiveIndex( index );
 }
 
 void QgsBrowserDockWidget::splitterMoved()
 {
-  QList<int> sizes = mSplitter->sizes();
-  float total = sizes.value( 0 ) + sizes.value( 1 );
-  mPropertiesWidgetHeight = total > 0 ? sizes.value( 1 ) / total : 0;
 }
