+++ src/ui/qgsbrowserwidgetbase.ui
@@ -1,19 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>QgsBrowserDockWidgetBase</class>
- <widget class="QgsDockWidget" name="QgsBrowserDockWidgetBase">
+ <class>QgsBrowserWidgetBase</class>
+ <widget class="QgsPanelWidget" name="QgsBrowserWidgetBase">
   <property name="geometry">
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>216</width>
-    <height>138</height>
+    <width>256</width>
+    <height>187</height>
    </rect>
   </property>
   <property name="windowTitle">
    <string>Browser</string>
   </property>
-  <widget class="QWidget" name="mContents">
    <layout class="QVBoxLayout" name="verticalLayout">
     <property name="spacing">
      <number>6</number>
@@ -189,7 +188,6 @@
      </widget>
     </item>
    </layout>
-  </widget>
   <action name="mActionAddLayers">
    <property name="icon">
     <iconset resource="../../images/images.qrc">
@@ -264,10 +262,9 @@
    <header>qgsfilterlineedit.h</header>
   </customwidget>
   <customwidget>
-   <class>QgsDockWidget</class>
-   <extends>QDockWidget</extends>
-   <header>qgsdockwidget.h</header>
-   <container>1</container>
+   <class>QgsPanelWidget</class>
+   <extends>QWidget</extends>
+   <header>qgspanelwidget.h</header>
   </customwidget>
  </customwidgets>
  <resources>
